+++ src/commands/setAvailabilityStatus.ts
@@ -1,8 +1,17 @@
 import {Availability} from '@wireapp/protocol-messaging';
 import {Account} from '@wireapp/core';
 import {APIClient} from '@wireapp/api-client';
-import {ClientType} from '@wireapp/api-client/src/client/';
+import {ClientClassification, ClientType} from '@wireapp/api-client/src/client/';
 import prompts from 'prompts';
+import * as path from 'path';
+import * as fs from 'fs';
+
+const defaultPackageJsonPath = path.join(__dirname, '../package.json');
+const packageJsonPath = fs.existsSync(defaultPackageJsonPath)
+  ? defaultPackageJsonPath
+  : path.join(__dirname, '../../package.json');
+
+const pkg = require(packageJsonPath);
 
 import {CommonOptions} from '../CommonOptions';
 import {getLogger, getBackendURL, getEmailAddress, getPassword, getWebSocketURL} from '../util';
@@ -11,7 +20,7 @@ const logger = getLogger('set-availability');
 
 export interface SetAvailabilityStatusOptions extends CommonOptions {
   defaultWebSocketURL: string;
-  statusType?: Availability.Type;
+  statusType?: Availability.Type | string | number;
   webSocketURL?: string;
 }
 
@@ -29,7 +38,36 @@ export async function setAvailabilityStatus({
   webSocketURL ||= await getWebSocketURL(defaultWebSocketURL);
   emailAddress ||= await getEmailAddress();
   password ||= await getPassword();
-  statusType ??= (
+
+  let newStatusType: Availability.Type;
+
+  if (typeof statusType === 'string' && !isNaN(Number(statusType.trim()))) {
+    newStatusType = Number(statusType.trim());
+  } else if (typeof statusType === 'string' && statusType.trim() !== '') {
+    switch (statusType.toUpperCase()) {
+      case 'NONE': {
+        newStatusType = Availability.Type.NONE;
+        break;
+      }
+      case 'AVAILABLE': {
+        newStatusType = Availability.Type.AVAILABLE;
+        break;
+      }
+      case 'AWAY': {
+        newStatusType = Availability.Type.AWAY;
+        break;
+      }
+      case 'BUSY': {
+        newStatusType = Availability.Type.BUSY;
+        break;
+      }
+      default: {
+        console.warn(`Invalid status type "${statusType}" set.`);
+      }
+    }
+  }
+
+  newStatusType ??= (
     await prompts(
       {
         choices: [
@@ -42,12 +80,12 @@ export async function setAvailabilityStatus({
             value: Availability.Type.AVAILABLE,
           },
           {
-            title: 'Busy',
-            value: Availability.Type.BUSY,
-          },
-          {
             title: 'Away',
             value: Availability.Type.AWAY,
+          },
+          {
+            title: 'Busy',
+            value: Availability.Type.BUSY,
           },
         ],
         message: 'Which status would you like to set?',
@@ -69,7 +107,11 @@ export async function setAvailabilityStatus({
   const account = new Account(apiClient);
 
   logger.info('Logging in ...');
-  await account.login({clientType: ClientType.TEMPORARY, email: emailAddress, password});
+  await account.login({clientType: ClientType.TEMPORARY, email: emailAddress, password}, undefined, {
+    classification: ClientClassification.DESKTOP,
+    cookieLabel: 'default',
+    model: `${pkg.name} ${pkg.version}`,
+  });
 
   const {team: teamId} = await account.service!.self.getSelf();
 
@@ -80,7 +122,7 @@ export async function setAvailabilityStatus({
   logger.info('Setting availability status ...');
 
   if (!dryRun) {
-    await account.service!.user.setAvailability(teamId, statusType!);
+    await account.service!.user.setAvailability(teamId, newStatusType);
   }
 
   logger.info('Logging out ...');